apps/api/requirements.txt
fastapi[standard]==0.115.0
sqlalchemy==2.0.36
pymysql==1.1.1
google-auth==2.36.0